requirements.txt
Django~=4.2.8
django-environ~=0.11.2
psycopg2~=2.9.9
djangorestframework~=3.14.0
django-rest-swagger~=2.2.0
